HSEARCH-2191 Remove feature packs from the distribution package
Feature packs cannot be consumed as local files with the currently
available tooling: one needs to refer to Maven artifacts available in a
repository. Thus making the feature packs available through direct
download is useless.

Also, even if it was possible, we should probably allow users to
download the feature-packs we depend on, and their transitive
dependencies. That would get quite complex quite fast, because we
currently depend on wildfly-feature-pack, which basically is the full
WildFly server, and that one has a *lot* of dependencies.

For now, we will keep the previous commits fixing problems related to
feature packs in the distribution package, and remove the feature packs
in this separate commit.
This will allow us to easily revert the decision to not make the feature
packs available through direct download, if we end up changing our minds.